+++ src/main/java/product_store/order/OrderOut.java
@@ -1,16 +1,14 @@
 package product_store.order;
 
 import lombok.Builder;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Builder
 public record OrderOut(
-
     String id,
     String date,
     List<OrderItemOut> items,
-    Float total
-
-) {
-    
-}
+    BigDecimal total
+) {}
